지도완성

+++ src/main/java/com/green/airline/repository/model/Airport.java
@@ -1,7 +1,5 @@
 package com.green.airline.repository.model;
 
-import com.green.airline.utils.DecimalPoint;
-
 import lombok.Data;
 
 // 공항
@@ -11,9 +9,7 @@ public class Airport {
 	private Integer id;
 	private String region;
 	private String name;
-	private float latitude;
-	//String formattedLatitude = DecimalPoint.formatDouble(latitude);
-	private float longitude;
-	//String formattedLongitude = DecimalPoint.formatDouble(longitude);
+	private double latitude;
+	private double longitude;
 	
 }
